refactor: tidy the components | SCRUM-313
