+++ src/app/ui/emailService/email-service-recipients-table.tsx
@@ -90,9 +90,6 @@ export const RecipientTable: React.FC<RecipientTableProps> = ({
   customColumns,
   recipients,
   setRecipients,
-  onAddColumn,
-  onAddRecipient,
-  onColumnValueChange,
   onDeleteColumn,
   onReorderColumns,
   onAddColumnWithName,
